app/requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Pillow==10.2.0
boto3==1.28.57
botocore==1.31.57
python-dotenv==1.0.1
